==> RNA/scripts/03_Br.mbio.e3.v3.0anno.DEG.2022.Rmd <==
---
title: "03_Br.mbio.e3.v3.0anno.DEG.2022"
author: "Kazu"
date: "2/4/2022"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
library(knitr)
library(GGally)
#library(d3heatmap)
#library(flexdashboard)
library(edgeR);library(tidyverse)
library(readr);library(readxl)
# library(UpSetR)
library(ggVennDiagram) # for VennDiagram
#url <- "http://datasets.flowingdata.com/ppg2008.csv"
#nba_players <- read.csv(url, row.names = 1)
```


# scatter plot exp1 vs exp3 (copied from "06_d_cor_permutation.Rd")
For each experiment, replace gene not meeting the threshold with "NAs".  This will allow GGpairs to function correctly and also will allow the permutation approach for p-values to work in GGpairs
## this part is under construction ### (Oct 16, 2021) ##

# prep for e1 and e3 data
### prep
```{r echo=TRUE, error=TRUE, message=FALSE, warning=FALSE}
# all model including subset data
DEG.objs.e1.e3.v3.0annotation.all <- list.files(path=file.path("RNA","output","FULLtoptag"),
                                           pattern="(^e1|^e3)(\\.)(leaf|root)(\\.)(.+)(v3.0anno)(.csv)") 
DEG.objs.e1.e3.v3.0annotation.all # 24 csv files. correct? (102121)
# read csv file (only full model)
DEG.count.list.e1.e3.v3.0annotation.all <- lapply(DEG.objs.e1.e3.v3.0annotation.all, function(x) read_csv(paste(file.path("RNA","output","FULLtoptag"),"/",x,sep="")))
names(DEG.count.list.e1.e3.v3.0annotation.all) <- gsub(".csv","",DEG.objs.e1.e3.v3.0annotation.all)
```

Overlap between DEGs (Oct 28, 2021) 
```{r error=TRUE}
## JM Added the line below on January 2025.  Not sure if this is correct, but otherwise the object does not exist.
DEG.count.list.DEG.objs.e1and3.add.trt.v3.0annotation <- DEG.count.list.e1.e3.v3.0annotation.all 

## JM Added the line below on January 2025.  Needed to get the names in the list to match the code below.
names(DEG.count.list.DEG.objs.e1and3.add.trt.v3.0annotation) <- names(DEG.count.list.DEG.objs.e1and3.add.trt.v3.0annotation) %>% str_remove(fixed(".v3.0anno"))


# FDR < 0.1
DEG.count.list.DEG.objs.e1and3.add.trt.v3.0annotation.FDR0.1 <- DEG.count.list.DEG.objs.e1and3.add.trt.v3.0annotation %>% map(.,~filter(.x,FDR < 0.1))
names(DEG.count.list.DEG.objs.e1and3.add.trt.v3.0annotation)
# 
DEG.count.list.DEG.objs.e1and3.add.trt.v3.0annotation.FDR0.1[["e3.root.trtlive.DEGs.add.rD.rCR"]] %>% View() # OK? 4975 genes, which is different from one in slide 14...
```


# combine exp1 and exp3
## root
```{r}
# DEGs of interest
e1.root.trtlive.DEGs <- DEG.count.list.DEG.objs.e1and3.add.trt.v3.0annotation.FDR0.1[["e1.root.trt5E_live.DEGs.add.rD.rFPsc.NB"]] %>% dplyr::select("genes","AGI","logFC","FDR") %>% drop_na(AGI) %>% dplyr::rename(`original root`=logFC,exp1root_FDR=FDR) # 63 genes # now 61
e3.root.cr.trtlive.DEGs <- DEG.count.list.DEG.objs.e1and3.add.trt.v3.0annotation.FDR0.1[["e3.root.trtlive.DEGs.add.rD.rCR"]] %>% dplyr::select("genes","AGI","logFC","FDR") %>% drop_na(AGI) %>% dplyr::rename(`validation root`=logFC,exp3.root_FDR=FDR)   # 4975 genes # now 4923
e1.leaf.trtlive.DEGs <- DEG.count.list.DEG.objs.e1and3.add.trt.v3.0annotation.FDR0.1[["e1.leaf.trt5E_live.DEGs.add.rD.rFPsc.NB"]] %>% dplyr::select("genes","AGI","logFC","FDR") %>% drop_na(AGI) %>% dplyr::rename(`original leaf`=logFC,exp1.leaf_FDR=FDR)  # 3729 genes # now 3694
e3.leaf.cr.trtlive.DEGs <- DEG.count.list.DEG.objs.e1and3.add.trt.v3.0annotation.FDR0.1[["e3.leaf.trtlive.DEGs.add.rD.rCR"]]  %>% dplyr::select("genes","AGI","logFC","FDR") %>% drop_na(AGI) %>% dplyr::rename(`validation leaf`=logFC,exp3.leaf_FDR=FDR)  # 319 genes # now 317

# combine e1 adn e3 (root): original ggpair()
root.exp1and3.FDR0.1.plot <- e1.root.trtlive.DEGs %>% inner_join(e3.root.cr.trtlive.DEGs,by="genes") %>% GGally::ggpairs(columns=c(3,6),lower = list(continuous = wrap("points", alpha = 0.2)))
# combine e1 adn e3 (leaf): original ggpair()
leaf.exp1and3.FDR0.1.plot <- e1.leaf.trtlive.DEGs %>% inner_join(e3.leaf.cr.trtlive.DEGs,by="genes") %>% GGally::ggpairs(columns=c(3,6),lower = list(continuous = wrap("points", alpha = 0.2)))


##### JM Janyary 2025: Dont `drop_na(AGI).`..keep all genes for correlation plots, etc.

# DEGs of interest
e1.root.trtlive.DEGs <- DEG.count.list.DEG.objs.e1and3.add.trt.v3.0annotation.FDR0.1[["e1.root.trt5E_live.DEGs.add.rD.rFPsc.NB"]] %>% dplyr::select("genes","AGI","logFC","FDR") %>% dplyr::rename(`original root`=logFC,exp1root_FDR=FDR) # 63 genes 
e3.root.cr.trtlive.DEGs <- DEG.count.list.DEG.objs.e1and3.add.trt.v3.0annotation.FDR0.1[["e3.root.trtlive.DEGs.add.rD.rCR"]] %>% dplyr::select("genes","AGI","logFC","FDR") %>% dplyr::rename(`validation root`=logFC,exp3.root_FDR=FDR)   # 4975 genes 
e1.leaf.trtlive.DEGs <- DEG.count.list.DEG.objs.e1and3.add.trt.v3.0annotation.FDR0.1[["e1.leaf.trt5E_live.DEGs.add.rD.rFPsc.NB"]] %>% dplyr::select("genes","AGI","logFC","FDR") %>% dplyr::rename(`original leaf`=logFC,exp1.leaf_FDR=FDR)  # 3729 genes 
e3.leaf.cr.trtlive.DEGs <- DEG.count.list.DEG.objs.e1and3.add.trt.v3.0annotation.FDR0.1[["e3.leaf.trtlive.DEGs.add.rD.rCR"]]  %>% dplyr::select("genes","AGI","logFC","FDR") %>% dplyr::rename(`validation leaf`=logFC,exp3.leaf_FDR=FDR)  # 319 genes

# combine e1 and e3 (root): original ggpair()
root.exp1and3.FDR0.1.plot <- e1.root.trtlive.DEGs %>% inner_join(e3.root.cr.trtlive.DEGs,by="genes") %>% ggpairs(columns=c(3,6),lower = list(continuous = wrap("points", alpha = 0.2)))
# combine e1 and e3 (leaf): original ggpair()
leaf.exp1and3.FDR0.1.plot <- e1.leaf.trtlive.DEGs %>% inner_join(e3.leaf.cr.trtlive.DEGs,by="genes") %>% ggpairs(columns=c(3,6),lower = list(continuous = wrap("points", alpha = 0.2)))

```


# VennDiagram ggVennDiagram version (Oct, 2021)
```{r}
venn.list <- list(
  e1.root=DEG.count.list.DEG.objs.e1and3.add.trt.v3.0annotation.FDR0.1[["e1.root.trt5E_live.DEGs.add.rD.rFPsc.NB"]] %>% dplyr::select(genes) %>% as_vector(),
  e3.root.cr=DEG.count.list.DEG.objs.e1and3.add.trt.v3.0annotation.FDR0.1[["e3.root.trtlive.DEGs.add.rD.rCR"]]  %>% dplyr::select(genes) %>% as_vector(),
  e1.leaf=DEG.count.list.DEG.objs.e1and3.add.trt.v3.0annotation.FDR0.1[["e1.leaf.trt5E_live.DEGs.add.rD.rFPsc.NB"]] %>% dplyr::select(genes) %>% as_vector(),
  e3.leaf.cr=DEG.count.list.DEG.objs.e1and3.add.trt.v3.0annotation.FDR0.1[["e3.leaf.trtlive.DEGs.add.rD.rCR"]]  %>% dplyr::select(genes) %>% as_vector())
# default
ggVennDiagram(venn.list)
# modify
venn.ggplot <- ggVennDiagram(
  venn.list, label_alpha = 0,label="count", edge_size=1,
  label_size=3,
  set_size=3,
  set_color="black",
  category.names = c("      original","      validation","original     ", "validation       ")
  ) + #ggplot2::scale_fill_manual(values = c("white", "white", "white","white")) +
  ggplot2::scale_fill_gradient(low="white",high = "white")  +
 theme(legend.position = "none") + scale_color_manual(values = c("black", "black", "green","green")) 
# adding "root" and "leaf" texts in venn.ggplot with different color (eg. black and green)
text.data <- tibble(x=c(0.1,0.9),y=c(0.9,0.9),label=c("root","leaf"),col=c("black","green"))
# venn.ggplot + geom_text(data=text.data, aes(x=x, y=y,label=label,colour=factor(label)),size=10,inherit.aes = FALSE) # produce error # I gave up
venn.ggplot <- venn.ggplot + geom_text(data=text.data, aes(x=x, y=y,label=label),size=4)
# how to delete %??
ggsave(file.path("RNA","output","venn.ggplot2.png"),width=8,height=6)

# error (do not use)
# text.plot <- ggplot() + geom_text(data=text.data, aes(x=x, y=y,label=label,colour=factor(label)),size=10) 
# text.plot %+%  ggVennDiagram(
#   venn.list, label_alpha = 0,label="count", edge_size=1,
#   set_color="black",
#   category.names = c("original","validation","original     ", "validation       ")
#   ) + #ggplot2::scale_fill_manual(values = c("white", "white", "white","white")) +
#   ggplot2::scale_fill_gradient(low="white",high = "white")  +
#  theme(legend.position = "none") + scale_color_manual(values = c("black", "black", "green","green")) 
# 
# # error (do not use)
# venn.ggplot %+%  geom_text(data=text.data, aes(x=x, y=y,label=label,colour=factor(label)),size=10)
```




# simpler cor.test() version. No permutation (Nov 2, 2021)
```{r}
root.exp1and3.FDR0.1.data <- e1.root.trtlive.DEGs %>% inner_join(e3.root.cr.trtlive.DEGs,by="genes") %>% dplyr::select(c(3,6)) 
leaf.exp1and3.FDR0.1.data <- e1.leaf.trtlive.DEGs %>% inner_join(e3.leaf.cr.trtlive.DEGs,by="genes") %>% dplyr::select(c(3,6)) 
#perm.cor(data=root.exp1and3.FDR0.1.data,xname=`original root`,yname=`validation root`,xdat=`original root`,ydat=`validation root`) # does not work. Do I need to use permutation?
# root
root.cor.test.result <- cor.test(~`original root`+`validation root`, root.exp1and3.FDR0.1.data)
# leaf
leaf.cor.test.result <- cor.test(~`original leaf`+`validation leaf`, leaf.exp1and3.FDR0.1.data)
# summary
cor.summary <- tibble(
  x = c(-0.7,-0.7,-0.7,-0.7), y=c(0.95,0.72,1.7,1.3),
  value=c(paste("cor = ",round(root.cor.test.result$estimate,2),sep=""),
          paste("p.value = ", signif(root.cor.test.result$p.value,2),sep=""),
          paste("cor = ",round(leaf.cor.test.result$estimate,2),sep=""),
          paste("p.value = ",signif(leaf.cor.test.result$p.value,2),sep="")))

# root plot
root.exp1and3.FDR0.1.plot.s <- ggplot(root.exp1and3.FDR0.1.data) + geom_point(aes(x=`original root`,y=`validation root`,alpha=0.5),size=1) + theme(legend.position="none") + geom_text(data=cor.summary[1:2,],aes(label=value,x=x,y=y),size=4) + theme(axis.text=element_text(size=10))
# leaf plot
leaf.exp1and3.FDR0.1.plot.s <- ggplot(leaf.exp1and3.FDR0.1.data) + geom_point(aes(x=`original leaf`,y=`validation leaf`,alpha=0.5),size=1) + theme(legend.position="none") +  geom_text(data=cor.summary[3:4,],aes(label=value,x=x,y=y),size=4)+ theme(axis.text=element_text(size=10))
```

# combine
```{r}
library(cowplot)
#venn.ggplot <- venn.ggplot + theme(plot.margin=margin(t = 0, r = 0, b = 0, l = 100, unit = "pt"))
#library(ggplotify) # https://github.com/ggobi/ggally/issues/256
p.all <- cowplot::plot_grid(root.exp1and3.FDR0.1.plot.s, leaf.exp1and3.FDR0.1.plot.s, venn.ggplot, labels = c('A','B','C'), label_size = 10,ncol=2,rel_heights =c(1,1.2)) # ,rel_widths=c(1,1,2)
p.all <- p.all+ theme(plot.background = element_rect(fill="white")) # default background in plot C was gray.
ggsave2(p.all,filename=file.path("RNA","output","cor.plus.Venn3.png"),width=6.5,height=7.5,units="in") # width 6.5 (Feb 3, 2022)
ggsave2(p.all,filename=file.path("RNA","output","cor.plus.Venn3.pdf"),width=6.5,height=7.5,units="in")

```

# Cor.test for the union (as compared to the intersection) of DEGs
```{r}
# get expression levels, regardless of FDR
e1.root.trtlive.all <- DEG.count.list.DEG.objs.e1and3.add.trt.v3.0annotation[["e1.root.trt5E_live.DEGs.add.rD.rFPsc.NB"]] %>% dplyr::select("genes","AGI","logFC","FDR") %>% dplyr::rename(`original root`=logFC,exp1.root_FDR=FDR) 
e3.root.cr.trtlive.all <- DEG.count.list.DEG.objs.e1and3.add.trt.v3.0annotation[["e3.root.trtlive.DEGs.add.rD.rCR"]] %>% dplyr::select("genes","AGI","logFC","FDR") %>% dplyr::rename(`validation root`=logFC,exp3.root_FDR=FDR)

e1.leaf.trtlive.all <- DEG.count.list.DEG.objs.e1and3.add.trt.v3.0annotation[["e1.leaf.trt5E_live.DEGs.add.rD.rFPsc.NB"]] %>% dplyr::select("genes","AGI","logFC","FDR") %>% dplyr::rename(`original leaf`=logFC,exp1.leaf_FDR=FDR)  
e3.leaf.cr.trtlive.all <- DEG.count.list.DEG.objs.e1and3.add.trt.v3.0annotation[["e3.leaf.trtlive.DEGs.add.rD.rCR"]]  %>% dplyr::select("genes","AGI","logFC","FDR") %>% dplyr::rename(`validation leaf`=logFC,exp3.leaf_FDR=FDR)

root.exp1and3.FDR0.1.data.union <- full_join(e1.root.trtlive.all, e3.root.cr.trtlive.all, by = "genes") %>%
  filter(exp1.root_FDR < 0.1 | exp3.root_FDR < 0.1) %>%
  dplyr::select(`original root`, `validation root`)

leaf.exp1and3.FDR0.1.data.union <- full_join(e1.leaf.trtlive.all, e3.leaf.cr.trtlive.all, by = "genes") %>%
  filter(exp1.leaf_FDR < 0.1 | exp3.leaf_FDR < 0.1) %>%
  dplyr::select(`original leaf`, `validation leaf`)

# root
root.cor.test.result.union <- cor.test(~`original root`+`validation root`, root.exp1and3.FDR0.1.data.union)
# leaf
leaf.cor.test.result.union <- cor.test(~`original leaf`+`validation leaf`, leaf.exp1and3.FDR0.1.data.union)
# summary
cor.summary.union <- tibble(
  x = c(-0.7,-0.7,-0.7,-0.7), y=c(0.95,0.72,1.7,1.3),
  value=c(paste("cor = ",round(root.cor.test.result.union$estimate,2),sep=""),
          paste("p.value = ", signif(root.cor.test.result.union$p.value,2),sep=""),
          paste("cor = ",round(leaf.cor.test.result.union$estimate,2),sep=""),
          paste("p.value = ",signif(leaf.cor.test.result.union$p.value,2),sep="")))

# root plot
root.exp1and3.FDR0.1.plot.s.union <- ggplot(root.exp1and3.FDR0.1.data.union) + 
  geom_point(aes(x=`original root`,y=`validation root`,alpha=0.5),size=1) + 
  theme(legend.position="none") + 
  geom_text(data=cor.summary.union[1:2,],aes(label=value,x=x,y=y),size=4) + 
  theme(axis.text=element_text(size=10))
# leaf plot
leaf.exp1and3.FDR0.1.plot.s.union <- ggplot(leaf.exp1and3.FDR0.1.data.union) + 
  geom_point(aes(x=`original leaf`,y=`validation leaf`,alpha=0.5),size=1) + 
  theme(legend.position="none") + 
  geom_text(data=cor.summary.union[3:4,],aes(label=value,x=x,y=y),size=4) + 
  theme(axis.text=element_text(size=10))
```
